requirements.txt
virtual-lab==1.1.0
streamlit==1.48.0
python-dotenv==1.1.1
openai==1.99.3
duckduckgo-search==6.2.6
